+++ plugins/apel/src/auditor_apel_plugin/publish.py
@@ -6,13 +6,14 @@
 import logging
 from pyauditor import AuditorClientBuilder
 from datetime import datetime, timedelta, timezone
+import json
 import configparser
 import argparse
 import base64
 from time import sleep
 from auditor_apel_plugin.core import (
     get_token,
-    get_time_db,
+    get_time_json,
     get_report_time,
     get_start_time,
     create_summary_db,
@@ -21,93 +22,89 @@ from auditor_apel_plugin.core import (
     sign_msg,
     build_payload,
     send_payload,
-    update_time_db,
+    update_time_json,
     get_begin_previous_month,
     get_begin_current_month,
     create_sync_db,
     group_sync_db,
     create_sync,
     get_records,
-    check_sites_in_records,
 )
 
 
 def run(config, client):
     report_interval = config["intervals"].getint("report_interval")
+    sites_to_report = json.loads(config["site"].get("sites_to_report"))
     token = get_token(config)
     logging.debug(token)
 
     while True:
-        time_db_conn = get_time_db(config)
-        last_report_time = get_report_time(time_db_conn)
+        time_dict = get_time_json(config)
+        last_report_time = get_report_time(time_dict)
         current_time = datetime.now()
         time_since_report = (current_time - last_report_time).total_seconds()
 
         if time_since_report < report_interval:
             logging.info("Not enough time since last report")
-            time_db_conn.close()
             sleep(report_interval - time_since_report)
             continue
         else:
             logging.info("Enough time since last report, create new report")
 
-        start_time = get_start_time(time_db_conn)
-        logging.info(f"Getting records since {start_time}")
+        for site in sites_to_report.keys():
+            logging.info(f"Getting records for {site}")
 
-        records_summary = get_records(config, client, start_time, 30)
+            start_time = get_start_time(config, time_dict, site)
+            logging.info(f"Getting records since {start_time}")
 
-        if len(records_summary) == 0:
-            logging.info("No new records, do nothing for now")
-            time_db_conn.close()
-            logging.info(
-                "Next report scheduled for "
-                f"{datetime.now() + timedelta(seconds=report_interval)}"
+            records_summary = get_records(config, client, start_time, 30, site=site)
+
+            if len(records_summary) == 0:
+                logging.info(f"No new records for {site}")
+                continue
+
+            latest_stop_time = records_summary[-1].stop_time.replace(
+                tzinfo=timezone.utc
             )
-            sleep(report_interval)
-            continue
 
-        sites_to_report = check_sites_in_records(config, records_summary)
-        logging.info(f"Create reports for {sites_to_report}")
+            logging.debug(f"Latest stop time is {latest_stop_time}")
+            summary_db = create_summary_db(config, records_summary)
+            grouped_summary_list = group_summary_db(summary_db)
+            summary = create_summary(config, grouped_summary_list)
+            logging.debug(summary)
+            signed_summary = sign_msg(config, summary)
+            logging.debug(signed_summary)
+            encoded_summary = base64.b64encode(signed_summary).decode("utf-8")
+            logging.debug(encoded_summary)
+            payload_summary = build_payload(encoded_summary)
+            logging.debug(payload_summary)
+            post_summary = send_payload(config, token, payload_summary)
+            logging.debug(post_summary.status_code)
 
-        latest_stop_time = records_summary[-1].stop_time.replace(tzinfo=timezone.utc)
+            if current_time.day == 1:
+                begin_month = get_begin_previous_month(current_time)
+            else:
+                begin_month = get_begin_current_month(current_time)
 
-        logging.debug(f"Latest stop time is {latest_stop_time}")
-        summary_db = create_summary_db(config, records_summary)
-        grouped_summary_list = group_summary_db(summary_db)
-        summary = create_summary(config, grouped_summary_list)
-        logging.debug(summary)
-        signed_summary = sign_msg(config, summary)
-        logging.debug(signed_summary)
-        encoded_summary = base64.b64encode(signed_summary).decode("utf-8")
-        logging.debug(encoded_summary)
-        payload_summary = build_payload(encoded_summary)
-        logging.debug(payload_summary)
-        post_summary = send_payload(config, token, payload_summary)
-        logging.debug(post_summary.status_code)
+            records_sync = get_records(config, client, begin_month, 30, site=site)
+            sync_db = create_sync_db(config, records_sync)
+            grouped_sync_list = group_sync_db(sync_db)
+            sync = create_sync(grouped_sync_list)
+            logging.debug(sync)
+            signed_sync = sign_msg(config, sync)
+            logging.debug(signed_sync)
+            encoded_sync = base64.b64encode(signed_sync).decode("utf-8")
+            logging.debug(encoded_sync)
+            payload_sync = build_payload(encoded_sync)
+            logging.debug(payload_sync)
+            post_sync = send_payload(config, token, payload_sync)
+            logging.debug(post_sync.status_code)
 
-        if current_time.day == 1:
-            begin_month = get_begin_previous_month(current_time)
-        else:
-            begin_month = get_begin_current_month(current_time)
+            latest_report_time = datetime.now()
+            update_time_json(
+                config, time_dict, site, latest_stop_time, latest_report_time
+            )
 
-        records_sync = get_records(config, client, begin_month, 30)
-        sync_db = create_sync_db(config, records_sync)
-        grouped_sync_list = group_sync_db(sync_db)
-        sync = create_sync(grouped_sync_list)
-        logging.debug(sync)
-        signed_sync = sign_msg(config, sync)
-        logging.debug(signed_sync)
-        encoded_sync = base64.b64encode(signed_sync).decode("utf-8")
-        logging.debug(encoded_sync)
-        payload_sync = build_payload(encoded_sync)
-        logging.debug(payload_sync)
-        post_sync = send_payload(config, token, payload_sync)
-        logging.debug(post_sync.status_code)
-
-        latest_report_time = datetime.now()
-        update_time_db(time_db_conn, latest_stop_time.timestamp(), latest_report_time)
-
-        time_db_conn.close()
         logging.info(
             "Next report scheduled for "
             f"{datetime.now() + timedelta(seconds=report_interval)}"